refactor advection example to better demonstrate a driver and a package.  Also use preludes.
